feat: add router pages and layout
