+++ package.json
@@ -6,8 +6,6 @@
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@speechly/react-client": "0.0.6",
-    "@speechly/react-ui": "^1.2.4",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.1",
@@ -17,8 +15,8 @@
     "react": "^16.12.0",
     "react-chartjs-2": "^2.11.1",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "recharts": "^2.0.9",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
