+++ src/App.js
@@ -1,8 +1,22 @@
 import React from 'react';
-import Dashboard from './components/Dashboard';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import Dashboard from './pages/Dashboard';
+import Reports from './pages/Reports';
+import useStyles from './styles';
 
 const App = () => {
-  return <Dashboard />;
+  const classes = useStyles();
+
+  return (
+    <BrowserRouter>
+      <div className={classes.root}>
+        <Switch>
+          <Route path="/reports" component={Reports} />
+          <Route path="/" component={Dashboard} />
+        </Switch>
+      </div>
+    </BrowserRouter>
+  );
 };
 
 export default App;
