+++ src/components/Navbar/index.js
@@ -1,0 +1,3 @@
+import Navbar from './Navbar';
+
+export default Navbar;
