+++ src/components/Navbar/styles.js
@@ -1,0 +1,46 @@
+import { purple } from '@material-ui/core/colors';
+import { makeStyles } from '@material-ui/core/styles';
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles(theme => ({
+  toolbar: {
+    paddingRight: 24, // keep right padding when drawer closed
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  },
+  username: {
+    marginRight: theme.spacing(2),
+    [theme.breakpoints.down('xs')]: {
+      display: 'none',
+    },
+  },
+  purple: {
+    color: theme.palette.getContrastText(purple[500]),
+    backgroundColor: purple[500],
+  },
+  menuButton: {
+    marginRight: 36,
+  },
+  menuButtonHidden: {
+    display: 'none',
+  },
+  title: {
+    flexGrow: 1,
+  },
+}));
+
+export default useStyles;
