+++ src/components/Sidebar/Sidebar.js
@@ -1,0 +1,36 @@
+import { Divider, Drawer, IconButton, List } from '@material-ui/core';
+import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
+import clsx from 'clsx';
+import React from 'react';
+import { mainListItems, secondaryListItems } from './listItems';
+import useStyles from './styles';
+
+const Sidebar = ({ open, setOpen }) => {
+  const classes = useStyles();
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
+  return (
+    <Drawer
+      variant="permanent"
+      classes={{
+        paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
+      }}
+      open={open}
+    >
+      <div className={classes.toolbarIcon}>
+        <IconButton onClick={handleDrawerClose}>
+          <ChevronLeftIcon />
+        </IconButton>
+      </div>
+      <Divider />
+      <List>{mainListItems}</List>
+      <Divider />
+      <List>{secondaryListItems}</List>
+    </Drawer>
+  );
+};
+
+export default Sidebar;
