+++ src/layout/MainLayout/index.js
@@ -1,0 +1,3 @@
+import MainLayout from './MainLayout';
+
+export default MainLayout;
