+++ src/pages/Reports/styles.js
@@ -1,0 +1,18 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(theme => ({
+  paper: {
+    padding: theme.spacing(2),
+    display: 'flex',
+    overflow: 'auto',
+    flexDirection: 'column',
+  },
+  fixedHeight: {
+    height: 250,
+    [theme.breakpoints.down('sm')]: {
+      height: '100%',
+    },
+  },
+}));
+
+export default useStyles;
