+++ src/styles.js
@@ -1,6 +1,9 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 export default makeStyles(theme => ({
+  root: {
+    display: 'flex',
+  },
   main: {
     background: theme.palette.background.paper,
     [theme.breakpoints.up('sm')]: {
